Implemented end points, unit tests

+++ src/main/java/jp/co/axa/apidemo/services/EmployeeServiceImpl.java
@@ -1,8 +1,13 @@
 package jp.co.axa.apidemo.services;
 
+import jp.co.axa.apidemo.dto.EmployeeRequest;
 import jp.co.axa.apidemo.entities.Employee;
+import jp.co.axa.apidemo.error.EmployeeNotFoundException;
 import jp.co.axa.apidemo.repositories.EmployeeRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.CachePut;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -18,25 +23,46 @@ public class EmployeeServiceImpl implements EmployeeService{
         this.employeeRepository = employeeRepository;
     }
 
+    @Cacheable(cacheNames = "employees")
     public List<Employee> retrieveEmployees() {
         List<Employee> employees = employeeRepository.findAll();
         return employees;
     }
 
-    public Employee getEmployee(Long employeeId) {
+    @Cacheable(cacheNames = "employees", key = "#employeeId")
+    public Employee getEmployee(Long employeeId) throws EmployeeNotFoundException {
         Optional<Employee> optEmp = employeeRepository.findById(employeeId);
+        if(!optEmp.isPresent()) {
+            throw new EmployeeNotFoundException("No Employee details found for id "+ employeeId);
+        }
         return optEmp.get();
     }
 
-    public void saveEmployee(Employee employee){
-        employeeRepository.save(employee);
+    public Employee saveEmployee(EmployeeRequest request){
+        Employee employee = Employee.
+                build( 0L, request.getName(), request.getSalary(),
+                        request.getDepartment());
+        return employeeRepository.save(employee);
     }
 
-    public void deleteEmployee(Long employeeId){
+    @CacheEvict(cacheNames =  "employees", key = "#employeeId")
+    public void deleteEmployee(Long employeeId) throws EmployeeNotFoundException{
+        Optional<Employee> optEmp = employeeRepository.findById(employeeId);
+        if(!optEmp.isPresent()) {
+            throw new EmployeeNotFoundException("No Employee details found for id "+ employeeId);
+        }
         employeeRepository.deleteById(employeeId);
     }
 
-    public void updateEmployee(Employee employee) {
+    @CachePut(cacheNames = "employees", key = "#employeeId")
+    public void updateEmployee(EmployeeRequest request, Long employeeId) throws EmployeeNotFoundException {
+        Optional<Employee> optEmp = employeeRepository.findById(employeeId);
+        if(!optEmp.isPresent()) {
+            throw new EmployeeNotFoundException("No Employee details found for id "+ employeeId);
+        }
+        Employee employee = Employee.build(employeeId,request.getName(), request.getSalary(),
+                request.getDepartment());
+
         employeeRepository.save(employee);
     }
 }
